payroll/requirements.txt
asgiref==3.5.1
Django==3.2.13
django-environ==0.4.5
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.2.0
psycopg2-binary==2.8.6